various tabs etc

+++ Opcodes/spectra.c
@@ -123,8 +123,8 @@ int spectset(CSOUND *csound, SPECTRUM *p)
       csound->Message(csound, Str("\thigh cps %7.1f\n\t low cps %7.1f\n"),
                               hicps, locps);
 
-      basfrq = hicps/2.0;                          /* oct below retuned top */
-      frqmlt = pow((double)2.0,(double)1.0/nfreqs);  /* nfreq interval mult */
+      basfrq = hicps/2.0;                     /* oct below retuned top */
+      frqmlt = pow(2.0,(double)1.0/nfreqs);   /* nfreq interval mult */
       Qfactor = Q * dwnp->srate;
       curfrq = basfrq;
       for (sumk=0,wsizp=p->winlen,woffp=p->offset,n=nfreqs; n--; ) {
